requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cssselect==1.1.0
dnspython==2.2.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.6.0
idna==2.10
jieba3k==0.35.1
joblib==1.1.0
lxml==4.8.0
newscatcherapi==0.4.0
newspaper3k==0.2.8
nltk==3.7
Pillow==9.0.1
pymongo==4.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
PyYAML==6.0
regex==2022.1.18
requests==2.24.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
tinysegmenter==0.3
tldextract==3.2.0
tqdm==4.62.3
urllib3==1.25.11
